Add files via upload
The lexycalAnalizer.py does the first part of the compilation process. 
It reads the input file character by character and separate it into lexemes that are turned into tokens of the form 
<lexeme_class, tokenValue>.
Then creates the symbol table where information about the tokens is stored. Reserved keywords, identifiers and user strings
are the lexemes inside the symbol tables. The information about them is:
* Lexeme name
* Amount of memory asigned
* Memory direction of the variables
* Line in which was found for first time
* Data type returned by the function
* Arguments that recieve the function

Then the tokens and the symbol table are delivered to the parser.